+++ gloabl.d.ts
@@ -1,0 +1,4 @@
+declare const __filename__: string;
+declare const __dirname__: string;
+declare const __version__: string;
+declare const __cwd__: string;