+++ packages/core/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 -   修改一些问题
 -   Updated dependencies
-    -   @lania/templates@0.0.3
+    -   @lania-cli/templates@0.0.3
 
 ## 0.0.2
 
@@ -14,4 +14,4 @@
 
 -   发布第一个版本
 -   Updated dependencies
-    -   @lania/templates@0.0.2
+    -   @lania-cli/templates@0.0.2